Move all strategies to their own file
