+++ tests/unit_tests/test_matchmaker_queue_algorithm.py
@@ -9,40 +9,8 @@ from hypothesis import strategies as st
 from server import config
 from server.matchmaker import Search, algorithm
 from server.rating import RatingType
-from tests.conftest import make_player
 
-
-@st.composite
-def st_players(draw, name="Player"):
-    """Strategy for generating Player objects"""
-    return make_player(
-        ladder_rating=(draw(st.floats(0, 2500)), draw(st.floats(1, 500))),
-        ladder_games=draw(st.integers(0, 1000)),
-        login=name
-    )
-
-
-@st.composite
-def st_searches(draw, num_players=1):
-    """Strategy for generating Search objects"""
-    return Search([
-        draw(st_players(f"p{i}")) for i in range(num_players)
-    ])
-
-
-@st.composite
-def st_searches_list(draw, min_players=1, max_players=10, max_size=30):
-    """Strategy for generating a list of Search objects"""
-    return draw(
-        st.lists(
-            st_searches(
-                num_players=draw(
-                    st.integers(min_value=min_players, max_value=max_players)
-                )
-            ),
-            max_size=max_size
-        )
-    )
+from .strategies import st_searches_list
 
 
 @pytest.fixture(scope="session")
